Create README.md
Brain-Tumor-MRI-Classification
In this Project I've used CNN to perform Image Classification on the Brain Tumor dataset. Since this dataset is small, if we train a neural network to it, it won't really give us a good result. Therefore I'm going to use the concept of Transfer Learning to train the model to get really accurate results